# Make performance more consistent by sorting dims before passing to `mean`, `min`...