@jsignell 2025-08-06
I opened this as an issue on numbagg: https://github.com/numbagg/numbagg/issues/402